requirements.txt
django==2.0.7
djangorestframework==3.8.2
mysqlclient==1.3.13
requests=2.19.1
Pillow=5.2.0
pyquery=1.4.0